Factoring orderings out to separate file and rename weightlex -> wlex, etc.
